Updating .gitignore, Comment added. main.m file added
Added a *.bak to .gitignore. This is the extension for backup files generated by texmaker that is use for writing the manuscript of this project. Added a short comment to the manuscript that was discussed with Nicola in a meeting. Finally, added a main.m file from which I will reflow or re-create the project workflow so it starts at this function. The reason behind this change is that given the previous state of the project, I did not know which script is supposed to first. Adding a main.m file insure that is the entry point for the project scripts