Initial commit
This is a survey of the performance of solving the eigen decomposition problem for a 2d or 3d image. The results suggest that a for-loop (or a parallel for-loop) are the most performant methods in this implementation. What is thus surprising is how fast the computations were in this script compared to using the fuction mshff. I will re-evaluate the file mshff.m to see why is it not as performant.